requirements.txt
pyserial
XBee